Move common load operations from topic.yaml.Loader to core.yaml.MigrationLoader
