+++ src/main/java/com/github/kattlo/topic/yaml/Loader.java
@@ -1,18 +1,15 @@
 package com.github.kattlo.topic.yaml;
 
-import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileReader;
 import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.Comparator;
 import java.util.Optional;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import java.util.stream.Stream;
 
 import com.github.kattlo.core.exception.LoadException;
+import com.github.kattlo.core.yaml.MigrationLoader;
 
 import org.mapstruct.factory.Mappers;
 import org.yaml.snakeyaml.Yaml;
@@ -37,68 +34,12 @@ public class Loader {
     private static final TopicOperationMapper MAPPER =
         Mappers.getMapper(TopicOperationMapper.class);
 
-    public static final Pattern VERSION_PATTERN =
-        Pattern.compile("(v[0-9]{4})");
-
-    public static final Pattern VERSION_NUMBER_PATTERN =
-        Pattern.compile("([0-9]{4})");
-
-    public static final Pattern FILE_NAME_PATTERN =
-        Pattern.compile("v[0-9]{4}_[\\w\\-]{0,246}\\.ya?ml");
-
-    public static final Pattern FILE_EXT_PATTERN =
-        Pattern.compile(".*\\.ya?ml");
-
-    static Stream<Path> list(final Path directory) throws IOException {
-        return Files.list(directory)
-            .filter(f ->
-                FILE_EXT_PATTERN.matcher(
-                    f.getFileName().toString())
-                        .matches());
-    }
-
-    public static Stream<Path> list(final File directory) throws IOException {
-        return list(Path.of(directory.getAbsolutePath()));
-    }
-
-    /**
-     * @throws IllegalArgumentException When the file name does not follow the pattern {@link #FILE_NAME_PATTERN}
-     */
-    public static void matches(Path file) {
-        if(!FILE_NAME_PATTERN.matcher(file.getFileName().toString()).matches()){
-            throw new IllegalArgumentException(file.getFileName().toString());
-        }
-    }
-
-    public static Optional<String> versionOf(Path file){
-
-        final Matcher m =
-            VERSION_PATTERN.matcher(file.getFileName().toString());
-
-        return
-            Optional.of(m.find())
-                .filter(found -> found)
-                .map(f -> m.group());
-
-    }
-
-    public static Optional<String> versionNumberOf(String version){
-
-        final Matcher m =
-            VERSION_NUMBER_PATTERN.matcher(version);
-
-        return
-            Optional.of(m.find())
-                .filter(found -> found)
-                .map(f -> m.group());
-    }
-
     /**
      * @throws LoadException When any problem happens to load the file
      * @throws IllegalArgumentException When the file name does not follow the pattern {@link #FILE_NAME_PATTERN}
      */
     public static Model load(Path file) {
-        matches(file);
+        MigrationLoader.matches(file);
 
         try {
             return YAML.load(new FileReader(file.toFile()));
@@ -117,7 +58,7 @@ public class Loader {
             final Path directory) throws IOException {
 
         return
-            list(directory)
+            MigrationLoader.list(directory)
                 .map(file ->
                     MAPPER.map(load(file), file))
                 .filter(o -> o.getTopic().equals(topic))
@@ -126,22 +67,14 @@ public class Loader {
                 .findFirst();
     }
 
-    static boolean greater(Path file, String currentVersion){
-        return
-            versionOf(file)
-                .filter(version -> version.compareTo(currentVersion) > 0)
-                .map(v -> Boolean.TRUE)
-                .orElse(Boolean.FALSE);
-    }
-
     public static Stream<TopicOperation> newer(
             final String currentVersion,
             final String topic,
             final Path directory) throws IOException {
 
         return
-            list(directory)
-                .filter(file -> greater(file, currentVersion))
+            MigrationLoader.list(directory)
+                .filter(file -> MigrationLoader.greater(file, currentVersion))
                 .map(file ->
                     MAPPER.map(load(file), file))
                 .filter(o -> o.getTopic().equals(topic))
@@ -154,7 +87,7 @@ public class Loader {
         final Path directory) throws IOException {
 
         return
-            list(directory)
+            MigrationLoader.list(directory)
                 .map(file -> MAPPER.map(load(file), file))
                 .filter(o -> o.getTopic().equals(topic))
                 .sorted(Comparator.comparing(TopicOperation::getVersion))
